requirements.txt
affine==2.4.0
aiofiles==23.1.0
anyio==3.6.2
argcomplete==2.0.0
attrs==22.2.0
boto3==1.26.78
botocore==1.29.78
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cycler==0.11.0
dacite==1.7.0
dbfread==2.0.7
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
eventlet==0.33.3
fiftyone==0.19.1
fiftyone-brain==0.10.0
fiftyone-db==0.4.0
fonttools==4.38.0
future==0.18.3
glob2==0.7
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.25.1
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
mongoengine==0.24.2
motor==3.1.1
ndjson==0.3.1
networkx==3.0
numpy==1.24.2
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
patool==1.12
Pillow==9.4.0
plotly==5.13.0
pprintpp==0.4.0
priority==2.0.0
psutil==5.9.4
pycocotools==2.0.6
pymongo==4.3.3
pyparsing==3.0.9
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
rasterio==1.3.6
requests==2.28.2
retrying==1.3.4
rfc3986==1.5.0
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
shapely==2.0.1
six==1.16.0
-e git+https://github.com/Sigmundom/oblique-aerial-image-labeling.git@e96673fec057712cb5b77961a561535fd67b09cc#egg=Skrafoto
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.20.4
strawberry-graphql==0.138.1
tabulate==0.9.0
tenacity==8.2.1
threadpoolctl==3.1.0
tifffile==2023.2.3
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.14
utm==0.7.0
voxel51-eta==0.8.3
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.14.0
